% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridbr-package.R
\docType{package}
\name{gridbr}
\alias{gridbr}
\title{gridbr: Download IBGE statistical grid}
\description{
Build a standardized grid and merge with population census data.
}
\references{
\url{https://github.com/luisfelipebr/gridbr}
}
\author{
Luis Cunha \email{luisfelipebc@outlook.com}
}
