% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridbr_crs.R
\docType{data}
\name{gridbr_crs}
\alias{gridbr_crs}
\title{IBGE statistical grid CRS (with equal area projection)}
\format{
An object of class \code{crs} of length 2.
}
\usage{
gridbr_crs
}
\description{
IBGE statistical grid CRS (with equal area projection)
}
\keyword{datasets}
