% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridbr_urban.R
\docType{data}
\name{gridbr_urban}
\alias{gridbr_urban}
\title{A character vector with the id of IBGE statistical grid urban cells ("200M" cell size).}
\format{
A character vector with 4610350 elements.
}
\usage{
data(gridbr_urban)
}
\description{
A character vector with the id of IBGE statistical grid urban cells ("200M" cell size).
}
\keyword{datasets}
