% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridbr_download.R
\name{gridbr_download}
\alias{gridbr_download}
\title{Download Brazilian statistical grid with population census data.}
\usage{
gridbr_download(input, cellsize, census_data = TRUE, equal_area = FALSE)
}
\arguments{
\item{input}{object of class sf, sfc or sfg.}

\item{cellsize}{character. Target cell size. Must be one of: "500KM", "100KM", "50KM", "10KM", "5KM", "1KM" and "200M".}

\item{census_data}{logical. Set to FALSE if you don't want to include 2010 population census data.}

\item{equal_area}{logical. Set to TRUE if you want to use the original grid CRS, with an equal area projection.}
}
\value{
Brazilian statistical grid
}
\description{
Make Brazilian statistical grid covering (\code{input}) with 2010 population census data.
}
